requirements.txt
fastapi
typing
uvicorn
google-generativeai
markdown-it-py
numpy
openpyxl
pandas
pillow
PyPDF2
python-docx
python-dotenv
requests